core: Move conversion.createMetadata() to metadata.fromRemoteDoc()
