+++ test/unit/metadata.js
@@ -8,6 +8,7 @@ const path = require('path')
 const MetadataBuilders = require('../support/builders/metadata')
 const { onPlatform } = require('../support/helpers/platform')
 
+const metadata = require('../../core/metadata')
 const {
   assignId,
   assignMaxDate,
@@ -23,12 +24,82 @@ const {
   buildFile,
   invariants,
   upToDate
-} = require('../../core/metadata')
+} = metadata
+const { FILES_DOCTYPE } = require('../../core/remote/constants')
+const timestamp = require('../../core/timestamp')
 
 const { platform } = process
 
 describe('metadata', function () {
   const builders = new MetadataBuilders()
+
+  describe('.fromRemoteDoc()', () => {
+    it('builds the metadata for a remote file', () => {
+      let remoteDoc /*: RemoteDoc */ = {
+        _id: '12',
+        _rev: '34',
+        _type: FILES_DOCTYPE,
+        class: 'document',
+        dir_id: '56',
+        executable: false,
+        md5sum: 'N7UdGUp1E+RbVvZSTy1R8g==',
+        mime: 'test/html',
+        name: 'bar',
+        path: '/foo/bar',
+        size: '78',
+        tags: ['foo'],
+        type: 'file',
+        updated_at: timestamp.stringify(timestamp.build(2017, 9, 8, 7, 6, 5))
+      }
+      let doc /*: Metadata */ = metadata.fromRemoteDoc(remoteDoc)
+
+      should(doc).deepEqual({
+        md5sum: 'N7UdGUp1E+RbVvZSTy1R8g==',
+        class: 'document',
+        docType: 'file',
+        updated_at: '2017-09-08T07:06:05Z',
+        mime: 'test/html',
+        path: 'foo/bar',
+        remote: {
+          _id: '12',
+          _rev: '34'
+        },
+        size: 78,
+        tags: ['foo']
+      })
+
+      remoteDoc.executable = true
+      doc = metadata.fromRemoteDoc(remoteDoc)
+      should(doc.executable).equal(true)
+    })
+
+    it('builds the metadata for a remote dir', () => {
+      const remoteDoc /*: RemoteDoc */ = {
+        _id: '12',
+        _rev: '34',
+        _type: FILES_DOCTYPE,
+        dir_id: '56',
+        name: 'bar',
+        path: '/foo/bar',
+        tags: ['foo'],
+        type: 'directory',
+        updated_at: timestamp.stringify(timestamp.build(2017, 9, 8, 7, 6, 5))
+      }
+
+      const doc = metadata.fromRemoteDoc(remoteDoc)
+
+      should(doc).deepEqual({
+        docType: 'folder',
+        updated_at: '2017-09-08T07:06:05Z',
+        path: 'foo/bar',
+        remote: {
+          _id: '12',
+          _rev: '34'
+        },
+        tags: ['foo']
+      })
+    })
+  })
 
   describe('assignId', function () {
     it('is available', function () {
