+++ test/unit/remote/index.js
@@ -13,9 +13,9 @@ const should = require('should')
 const stream = require('stream')
 
 const { withContentLength } = require('../../../core/file_stream_provider')
-const conversion = require('../../../core/conversion')
 const checksumer = require('../../../core/local/checksumer')
-const { ensureValidPath } = require('../../../core/metadata')
+const metadata = require('../../../core/metadata')
+const { ensureValidPath } = metadata
 const Prep = require('../../../core/prep')
 const { Remote } = require('../../../core/remote')
 const { TRASH_DIR_ID } = require('../../../core/remote/constants')
@@ -66,7 +66,7 @@ describe('remote.Remote', function () {
         .contentType('image/jpeg').dataFromFile(fixture).create()
 
       should(binary.md5sum).equal(expectedChecksum)
-      const stream = await this.remote.createReadStreamAsync(conversion.createMetadata(binary))
+      const stream = await this.remote.createReadStreamAsync(metadata.fromRemoteDoc(binary))
       should.exist(stream)
       const checksum = crypto.createHash('md5')
       checksum.setEncoding('base64')
@@ -249,7 +249,7 @@ describe('remote.Remote', function () {
     it('does nothing when the folder already exists', async function () {
       const parentDir /*: RemoteDoc */ = await builders.remote.dir().create()
       const remoteDir /*: RemoteDoc */ = await builders.remote.dir().inDir(parentDir).create()
-      const doc /*: Metadata */ = _.merge({remote: undefined}, conversion.createMetadata(remoteDir))
+      const doc /*: Metadata */ = _.merge({remote: undefined}, metadata.fromRemoteDoc(remoteDir))
       ensureValidPath(doc)
 
       await this.remote.addFolderAsync(doc)
@@ -276,7 +276,7 @@ describe('remote.Remote', function () {
     describe('overwriteFileAsync', function () {
       it('overwrites the binary content', async function () {
         const created = await builders.remote.file().data('foo').timestamp(2015, 11, 16, 16, 12, 1).create()
-        const old = conversion.createMetadata(created)
+        const old = metadata.fromRemoteDoc(created)
         const doc /*: Metadata */ = _.defaults({
           _id: created._id,
           md5sum: 'N7UdGUp1E+RbVvZSTy1R8g==',
@@ -308,7 +308,7 @@ describe('remote.Remote', function () {
 
       it('throws an error if the checksum is invalid', async function () {
         const created = await builders.remote.file().data('foo').create()
-        const old = conversion.createMetadata(created)
+        const old = metadata.fromRemoteDoc(created)
         const doc = _.defaults({md5sum: 'Invalid///////////////=='}, old)
         this.remote.other = {
           createReadStreamAsync (localDoc) {
@@ -344,7 +344,7 @@ describe('remote.Remote', function () {
   describe('updateFileMetadataAsync', () => {
     it('makes the remote file executable when the local one was made too', async function () {
       const oldRemote = await builders.remote.file().executable(false).create()
-      const old = conversion.createMetadata(oldRemote)
+      const old = metadata.fromRemoteDoc(oldRemote)
       const doc = _.defaults({executable: true}, old)
 
       await this.remote.updateFileMetadataAsync(doc, old)
@@ -356,7 +356,7 @@ describe('remote.Remote', function () {
 
     it('makes the remote file non-executable when the local one is not anymore', async function () {
       const oldRemote = await builders.remote.file().executable(true).create()
-      const old = conversion.createMetadata(oldRemote)
+      const old = metadata.fromRemoteDoc(oldRemote)
       const doc = _.clone(old)
       delete doc.executable
 
@@ -412,7 +412,7 @@ describe('remote.Remote', function () {
         .name('old-name')
         .timestamp(2017, 11, 15, 8, 12, 9)
         .create()
-      const old /*: Metadata */ = conversion.createMetadata(created)
+      const old /*: Metadata */ = metadata.fromRemoteDoc(created)
       const newParentDir /*: RemoteDoc */ = await builders.remote.dir()
         .name('new-parent-dir')
         .inRootDir()
@@ -446,7 +446,7 @@ describe('remote.Remote', function () {
         .inDir(parentDir)
         .timestamp(2016, 1, 2, 3, 4, 5)
         .create()
-      const oldMetadata /*: Metadata */ = conversion.createMetadata(deletedDir)
+      const oldMetadata /*: Metadata */ = metadata.fromRemoteDoc(deletedDir)
       const newMetadata /*: Metadata */ = _.defaults({
         name: 'new-dir-name',
         path: path.normalize('parent-dir/new-dir-name')
@@ -475,7 +475,7 @@ describe('remote.Remote', function () {
           remote: undefined,
           updated_at: timestamp.stringify(timestamp.build(2015, 1, 1, 1, 1, 1))
         },
-        conversion.createMetadata(
+        metadata.fromRemoteDoc(
           builders.remote.dir().name('foo').build()
         )
       )
@@ -507,7 +507,7 @@ describe('remote.Remote', function () {
         .name('cat6.jpg')
         .data('meow')
         .create()
-      old = conversion.createMetadata(remoteDoc)
+      old = metadata.fromRemoteDoc(remoteDoc)
       doc = _.defaults({
         path: path.normalize('moved-to/cat7.jpg'),
         name: 'cat7.jpg',
@@ -637,7 +637,7 @@ describe('remote.Remote', function () {
   describe('trash', () => {
     it('moves the file or folder to the Cozy trash', async function () {
       const folder = await builders.remote.dir().create()
-      const doc = conversion.createMetadata(folder)
+      const doc = metadata.fromRemoteDoc(folder)
 
       await this.remote.trashAsync(doc)
 
@@ -647,7 +647,7 @@ describe('remote.Remote', function () {
 
     it('does nothing when file or folder does not exist anymore', async function () {
       const folder = await builders.remote.dir().build()
-      const doc = conversion.createMetadata(folder)
+      const doc = metadata.fromRemoteDoc(folder)
 
       await this.remote.trashAsync(doc)
 
@@ -659,7 +659,7 @@ describe('remote.Remote', function () {
   describe('deleteFolderAsync', () => {
     it('deletes permanently an empty folder', async function () {
       const folder = await builders.remote.dir().create()
-      const doc = conversion.createMetadata(folder)
+      const doc = metadata.fromRemoteDoc(folder)
 
       await this.remote.deleteFolderAsync(doc)
 
@@ -669,7 +669,7 @@ describe('remote.Remote', function () {
 
     it('trashes a non-empty folder', async function () {
       const dir = await builders.remote.dir().create()
-      const doc = conversion.createMetadata(dir)
+      const doc = metadata.fromRemoteDoc(dir)
       await builders.remote.dir().inDir(dir).create()
 
       await this.remote.deleteFolderAsync(doc)
@@ -680,7 +680,7 @@ describe('remote.Remote', function () {
 
     it('resolves when folder does not exist anymore', async function () {
       const dir = await builders.remote.dir().build()
-      const doc = conversion.createMetadata(dir)
+      const doc = metadata.fromRemoteDoc(dir)
 
       await this.remote.deleteFolderAsync(doc)
 
@@ -690,7 +690,7 @@ describe('remote.Remote', function () {
 
     it('resolves when folder is being deleted (race condition)', async function () {
       const dir = await builders.remote.dir().create()
-      const doc = conversion.createMetadata(dir)
+      const doc = metadata.fromRemoteDoc(dir)
       sinon.stub(this.remote.remoteCozy, 'isEmpty').callsFake(async (id) => {
         await cozy.files.destroyById(id)
         return true
@@ -705,19 +705,19 @@ describe('remote.Remote', function () {
 
     it('does not swallow trashing errors', async function () {
       const dir = await builders.remote.dir().trashed().create()
-      const doc = conversion.createMetadata(dir)
+      const doc = metadata.fromRemoteDoc(dir)
       await should(this.remote.deleteFolderAsync(doc)).be.rejected()
     })
 
     it('does not swallow emptiness check errors', async function () {
       const file = await builders.remote.file().create()
-      const doc = conversion.createMetadata(file)
+      const doc = metadata.fromRemoteDoc(file)
       await should(this.remote.deleteFolderAsync(doc)).be.rejected()
     })
 
     it('does not swallow destroy errors', async function () {
       const dir = await builders.remote.dir().create()
-      const doc = conversion.createMetadata(dir)
+      const doc = metadata.fromRemoteDoc(dir)
       sinon.stub(this.remote.remoteCozy, 'destroyById').rejects('whatever')
       await should(this.remote.deleteFolderAsync(doc)).be.rejected()
     })
@@ -732,17 +732,17 @@ describe('remote.Remote', function () {
       remote.dst.dir = await this.remote.remoteCozy.updateAttributesById(remote.src.dir._id, {name: 'dst-dir'})
       remote.dst.foo = await this.remote.remoteCozy.find(remote.src.foo._id)
 
-      const doc /*: Metadata */ = conversion.createMetadata(remote.src.foo)
+      const doc /*: Metadata */ = metadata.fromRemoteDoc(remote.src.foo)
       doc.path = 'dst-dir/foo' // File metadata was updated as part of the move
       await this.remote.assignNewRev(doc)
-      should(doc).deepEqual(conversion.createMetadata(remote.dst.foo))
+      should(doc).deepEqual(metadata.fromRemoteDoc(remote.dst.foo))
     })
   })
 
   describe('renameConflictingDocAsync', () => {
     it('renames the file/folder', async function () {
       const remoteDoc /*: RemoteDoc */ = await builders.remote.file().name('cat9').create()
-      const src /*: Metadata */ = conversion.createMetadata(remoteDoc)
+      const src /*: Metadata */ = metadata.fromRemoteDoc(remoteDoc)
       ensureValidPath(src)
       const newPath = 'cat9-conflict-2015-12-01T01:02:03Z.jpg'
       await this.remote.renameConflictingDocAsync(src, newPath)
